requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.2.0
beir==1.0.0
blis==0.7.11
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm @ file:///croot/comm_1709322850197/work
confection==0.1.5
cymem==2.0.8
datasets==2.20.0
debugpy @ file:///home/builder/ci_310/debugpy_1640789504635/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepspeed==0.14.4
dill==0.3.8
einops==0.8.0
elasticsearch==7.9.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
grpcio==1.64.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipykernel @ file:///croot/ipykernel_1671488378391/work
ipython @ file:///croot/ipython_1718287989724/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.4
joblib==1.4.2
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openrlhf @ file:///home/student2021/syq/OpenRLHF
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.11.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
preshed==3.0.9
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==4.25.3
psutil @ file:///home/builder/ci_310/psutil_1640792629460/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.0
pydantic_core==2.20.0
Pygments @ file:///croot/pygments_1684279966437/work
PyJWT==2.8.0
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytrec_eval==0.5
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///home/builder/ci_310/pyzmq_1640794096902/work
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge==1.0.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tornado @ file:///home/builder/ci_310/tornado_1640795392144/work
tqdm==4.66.4
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.42.3
triton==2.3.1
typer==0.12.3
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
wasabi==1.1.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zhipuai==2.1.3
